+++ src/main/java/kitchenpos/order/application/OrderValidator.java
@@ -1,0 +1,6 @@
+package kitchenpos.order.application;
+
+public interface OrderValidator {
+
+    void validateEmpty(Long orderId);
+}
