+++ src/test/java/kitchenpos/domain/OrderTest.java
@@ -1,32 +1,36 @@
 package kitchenpos.domain;
 
-import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.assertThat;
 
+import java.time.LocalDateTime;
 import kitchenpos.order.domain.Order;
-import kitchenpos.table.domain.OrderTable;
+import kitchenpos.order.domain.OrderStatus;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
 class OrderTest {
 
     @Test
-    void Order_객체를_생성한다() {
+    void 주문의_상태가_완료_상태가_아닌지_확인한다() {
         // given
-        final OrderTable orderTable = new OrderTable(null, 1L, 1, false);
+        final Order order = new Order(1L, OrderStatus.COOKING, LocalDateTime.now());
 
-        // when & then
-        assertThatCode(() -> Order.from(orderTable))
-                .doesNotThrowAnyException();
+        // when
+        final boolean actual = order.isCompletion();
+
+        // then
+        assertThat(actual).isFalse();
     }
 
     @Test
-    void OrderTable이_빈_상태이면_Order_객체를_생성할_수_없다() {
+    void 주문의_상태가_완료_상태가_맞는지_확인한다() {
         // given
-        final OrderTable orderTable = new OrderTable(null, 1L, 1, true);
+        final Order order = new Order(null, OrderStatus.COMPLETION, LocalDateTime.now());
 
-        // when & then
-        assertThatThrownBy(() -> Order.from(orderTable))
-                .isInstanceOf(IllegalArgumentException.class);
+        // when
+        final boolean actual = order.isCompletion();
+
+        // then
+        assertThat(actual).isTrue();
     }
 }
